% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughnet.R
\name{roughnet}
\alias{roughnet}
\title{Create a rough network}
\usage{
roughnet(
  g,
  roughness = c(1, 1),
  bowing = c(1, 1),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{g}{igraph object}

\item{roughness}{numeric vector for roughness of vertices and edges}

\item{bowing}{numeric vector for bowing of vertices and edges}

\item{width}{width}

\item{height}{height}

\item{elementId}{DOM id}
}
\description{
plot a network using rough.js
}
\details{
the function recognizes the following attributes
Vertex attributes (e.g. V(g)$shape):
\itemize{
\item \emph{shape} either "circle" or "rectangle"
\item \emph{fill} vertex fill color
\item \emph{color} vertex stroke color
\item \emph{fillstyle} one of "hachure", "solid", "zigzag", "cross-hatch", "dots", "sunburst", "dashed", "zigzag-line"
\item \emph{size} vertex size
}

Edge attributes (e.g. E(g)$color):
\itemize{
\item \emph{color} edge color
\item \emph{width} edge width
}

Default values are used if one of the attributes is not found.

The result of a roughnet call can be prointed to file with \code{save_roughnet()}

More details on roughjs can be found on https://github.com/rough-stuff/rough/wiki
}
