% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_roughnet.R
\name{save_roughnet}
\alias{save_roughnet}
\title{Save roughnet plot to file}
\usage{
save_roughnet(rnet, file)
}
\arguments{
\item{rnet}{result from calling the function \code{roughnet}}

\item{file}{filename}
}
\description{
Save roughnet plot to file
}
